drm/bridge: Add an ->atomic_check() hook
So that bridge drivers have a way to check/reject an atomic operation.
The drm_atomic_bridge_chain_check() (which is just a wrapper around
the ->atomic_check() hook) is called in place of
drm_bridge_chain_mode_fixup() (when ->atomic_check() is not implemented,
the core falls back to ->mode_fixup(), so the behavior should stay
the same for existing bridge drivers).

Signed-off-by: Boris Brezillon <boris.brezillon@collabora.com>
---
Laurent,

I haven't clarified what's allowed to be changed in
->mode_fixup()/->atomic_check() simply because I don't know what the
constraints are. Since the problem already exists with the
->mode_fixup() hook, and given the goal of this series is to support
bus_format negotation (which does not involve tweaking the display
mode), I'd prefer to delay these semantic adjustments if you're okay
with that.

Regards,

Boris

Changes in v2:
* Clarify the fact that ->atomic_check() is replacing ->mode_fixup()